+++ example/static_alloc_config.yaml
@@ -10,8 +10,18 @@
   labels:
     job_uniq_id: '0'
   allocations:
-    cpu_shares: 0.5
-    cpu_quota: 0.01
+    # cpu_shares: 0.5
+    # cpu_quota: 0.01
     rdt:
         name: 'be'
-        l3: 'L3:0=00002;1=00002'
+        l3: 'L3:0=0ffff;1=0ffff'
+
+- name: 1-9 tasks
+  labels:
+    job_uniq_id: '[123456789]'
+  allocations:
+    cpu_shares: 0.1
+    cpu_quota: 0.05
+    rdt:
+        name: 'onetoten'
+        l3: 'L3:0=00001;1=00001'
