+++ owca/components.py
@@ -16,17 +16,18 @@
 from typing import List
 import pkg_resources
 
+import owca.runners.allocation
+import owca.runners.detection
 from owca import config
 from owca import detectors
 from owca import allocators
 from owca import mesos
-from owca import runner
 from owca import storage
 
 
 def register_components(extra_components: List[str]):
-    config.register(runner.DetectionRunner)
-    config.register(runner.AllocationRunner)
+    config.register(owca.runners.detection.DetectionRunner)
+    config.register(owca.runners.allocation.AllocationRunner)
     config.register(mesos.MesosNode)
     config.register(storage.LogStorage)
     config.register(storage.KafkaStorage)
